Adding support to float matrix

+++ src/convertModel.jl
@@ -3,11 +3,11 @@ using LinearAlgebra
 function convert(model::Model)
 
     c = Array{Float64,1}()
-    b = Array{Int64,1}()
-    A = Array{Int64}(undef, 0, num_variables(model))
+    b = Array{Float64,1}()
+    A = Array{Float64}(undef, 0, num_variables(model))
     z = 0
 
-    restriction = reshape(zeros(Int64,num_variables(model)), 1, num_variables(model))
+    restriction = reshape(zeros(Float64,num_variables(model)), 1, num_variables(model))
 
     min_max = Int8(objective_sense(model))
 
@@ -43,7 +43,7 @@ function convert(model::Model)
     
     constraints = all_constraints(model, AffExpr, MOI.LessThan{Float64})
     for i=1:num_constraints(model, AffExpr, MOI.LessThan{Float64})
-        restriction = reshape(zeros(Int64,num_variables(model)), 1, num_variables(model))
+        restriction = reshape(zeros(Float64,num_variables(model)), 1, num_variables(model))
         for j=1:num_variables(model)
             restriction[j] = normalized_coefficient(constraints[i],variables[j])    
         end
@@ -51,20 +51,20 @@ function convert(model::Model)
         append!(b,normalized_rhs(constraints[i]))
         append!(c,0)
     end
-    A = hcat(A, Matrix{Int64}(I, num_constraints(model, AffExpr, MOI.LessThan{Float64}), num_constraints(model, AffExpr, MOI.LessThan{Float64})))
+    A = hcat(A, Matrix{Float64}(I, num_constraints(model, AffExpr, MOI.LessThan{Float64}), num_constraints(model, AffExpr, MOI.LessThan{Float64})))
     M = 1000*maximum(broadcast(abs, c))
     
     constraints = all_constraints(model, AffExpr, MOI.EqualTo{Float64})
     for i=1:num_constraints(model, AffExpr, MOI.EqualTo{Float64})
-        restriction = reshape(zeros(Int64,num_variables(model)), 1, num_variables(model))
+        restriction = reshape(zeros(Float64,num_variables(model)), 1, num_variables(model))
         append!(c,-M)
         append!(b,normalized_rhs(constraints[i]))
         for j=1:num_variables(model)
             restriction[j] = normalized_coefficient(constraints[i],variables[j])    
         end
-        restriction = hcat(restriction,transpose(zeros(Int64,size(A,2)-num_variables(model))))
+        restriction = hcat(restriction,transpose(zeros(Float64,size(A,2)-num_variables(model))))
         A = vcat(A,restriction)
-        A = hcat(A,[zeros(Int64,size(A,1)-1);1])
+        A = hcat(A,[zeros(Float64,size(A,1)-1);1])
         
         c -= c[end]*A[end, :]
         z += b[end]*M
@@ -73,16 +73,16 @@ function convert(model::Model)
     
     constraints = all_constraints(model, AffExpr, MOI.GreaterThan{Float64})
     for i=1:num_constraints(model, AffExpr, MOI.GreaterThan{Float64})
-        restriction = reshape(zeros(Int64,num_variables(model)), 1, num_variables(model))
+        restriction = reshape(zeros(Float64,num_variables(model)), 1, num_variables(model))
         append!(c,[0;-M])
         append!(b,normalized_rhs(constraints[i]))
         for j=1:num_variables(model)
             restriction[j] = normalized_coefficient(constraints[i],variables[j])    
         end
-        restriction = hcat(restriction,transpose(zeros(Int64,size(A,2)-num_variables(model))))
+        restriction = hcat(restriction,transpose(zeros(Float64,size(A,2)-num_variables(model))))
         A = vcat(A,restriction)
-        A = hcat(A,[zeros(Int64,size(A,1)-1);1])
-        A = hcat(A,[zeros(Int64,size(A,1)-1);-1])
+        A = hcat(A,[zeros(Float64,size(A,1)-1);1])
+        A = hcat(A,[zeros(Float64,size(A,1)-1);-1])
         c -= c[end-1]*A[end, :]
         z += b[end]*M
     end
